Add support for new Muzei API
