+++ app/src/main/AndroidManifest.xml
@@ -88,14 +88,16 @@
         </activity>
 
         <!-- Muzei Service -->
-        <service
+        <provider
             android:name=".MuzeiService"
+            android:authorities="${applicationId}.muzei"
             android:label="@string/app_name"
             android:description="@string/app_description"
             android:icon="@drawable/ic_muzei_icon"
-            tools:ignore="ExportedService">
+            android:exported="true"
+            android:permission="com.google.android.apps.muzei.api.ACCESS_PROVIDER">
             <intent-filter>
-                <action android:name="com.google.android.apps.muzei.api.MuzeiArtSource" />
+                <action android:name="com.google.android.apps.muzei.api.MuzeiArtProvider" />
             </intent-filter>
 
             <meta-data
@@ -104,7 +106,7 @@
             <meta-data
                 android:name="settingsActivity"
                 android:value="jahirfiquitiva.libs.frames.ui.activities.MuzeiSettingsActivity" />
-        </service>
+        </provider>
 
         <!-- Required for files access !!! DO NOT REMOVE !!! -->
         <provider
