+++ app/src/main/kotlin/jahirfiquitiva/apps/frames/sample/MuzeiService.kt
@@ -15,7 +15,7 @@
  */
 package jahirfiquitiva.apps.frames.sample
 
-import jahirfiquitiva.libs.frames.data.services.FramesArtSource
+import jahirfiquitiva.libs.frames.data.services.FramesArtProvider
 
 // TODO: Use a unique name instead of MyFramesMuzeiService
-class MuzeiService : FramesArtSource("MyFramesMuzeiService")
+class MuzeiService : FramesArtProvider()
